Added skip turn and spawn wall infront of base feature
Added skip turn and changed the upgrade for base getting bigger, to a wall spawns infront of the base  for protection instead. Also moved players money from being a game object in the scene heirarchy, to inside the player prefab, as a script called 'character currency'. Also nullified the tower upgrade script - will replace it with something else later