exit game + update database
MainScene has the object GameExiterTemporary which simulates a game ending.

While in the scene, pressing P will exit the game and return to the main menu. by calling the function exitGame(). To be replaced by when the game ends (change if condition from pressing P to detect if game ends.)

While in the scene, pressing B will add a win to the database by calling IncrementWinsC(). To be replaced by when the win screen appears for the user.